requirements.txt
argcomplete
argh
ccxt
configobj
dill
forwardable
icecream
mailer
python-box
retry
sympy
tabulate
tqdm
